Create mrawin.bat
This is a batch script to facilitate creating *.arc and *.rom files with mra.exe on Windows. To use it, put the *.bat file into the same folder as the mra executeable, add some *.mra files (they can also be in a subfolder called mra) and corresponding MAME *.zip files (they can also be in a subfolder called mame) and then run the *.bat file.